\name{NLDSD}
\alias{NLDSD}
\title{
Nonlinear Data Stream Generator
}
\description{
Abstract class for NLDSD (Nonlinear Data Stream Generator)
}
\details{
The ‘NLDSD’ class cannot be instantiated, but it serves as a abstract base class from which all NLDSD objects inherit.

‘NLDSD’ provides common functionality like ‘print()’, ‘plot()’, etc.

Note that calling ‘NLDSD()’ results in an error since it is an abstract classes.
}
\author{
Fausto Guzzo da Costa
}
\seealso{
\code{\link{getpoints}}
}
