\name{plot}
\alias{plot}
\alias{plot.NLDSD}
\title{
Plotting Nonlinear Data Stream Data
}
\description{
Method to plotting nonlinear data stream data.
}
\usage{
plot(nldsd, n=500)
}
\arguments{
  \item{nldsd}{The NLDSD object to be plotted.}
  \item{n}{number of observations taken from NLDSD to plot.}
}
\author{
    Fausto Guzzo da Costa
}
\seealso{
    \code{\link{NLDSD}}
}
\examples{
    # TODO
}
